agents/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
kafka-python==2.0.2
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2
scikit-learn==1.3.2
tensorflow==2.15.0

# Data sources
python-binance==1.0.19
tweepy==4.14.0
praw==7.7.1

# Monitoring
prometheus-client==0.19.0
